refactor: Extract OptionsForm and improve state management
- Create OptionsForm component for form logic separation
  - Handles react-hook-form integration
  - Manages saved state internally
  - Receives non-nullable Settings
- Refactor OptionsPage to focus on layout and loading
  - Remove isLoading prop (use initialSettings === undefined)
  - Use ComponentProps<typeof OptionsForm> for type safety
  - Delegate form rendering to OptionsForm
- Simplify OptionsApp
  - Remove redundant isLoading state
  - Clean up effect handler
- Improve type safety and DRY
  - OptionsPage props extend from OptionsForm props
  - Single source of truth for prop definitions
- Add OptionsForm stories for component testing
- Update OptionsPage stories (remove isLoading, simplify)

Benefits:
- Better separation of concerns (App/Page/Form layers)
- Improved type safety (OptionsForm always has Settings)
- Reduced redundancy (no duplicate state for loading)
- Better testability (components can be tested independently)
